60-day-bootcamp/day-16-37/whisper-finetune/requirements.txt
datasets[audio]
transformers
accelerate
evaluate
jiwer
tensorboard
gradio
torchaudio
torchvision
soundfile
librosa
huggingface-hub